+++ strings/strings_test.go
@@ -104,36 +104,6 @@ func TestSwapCase(t *testing.T) {
 	}
 }
 
-func TestRJust(t *testing.T) {
-	type args struct {
-		s        string
-		width    int
-		fillChar rune
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "TestRJust",
-			args: args{
-				s:        "0123",
-				width:    5,
-				fillChar: ' ',
-			},
-			want: " 0123",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := RJust(tt.args.s, tt.args.width, tt.args.fillChar); got != tt.want {
-				t.Errorf("RJust() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func TestLength(t *testing.T) {
 	type args struct {
 		s string
